backend/requirements.txt
crewai==0.51.0
crewai-tools==0.8.3
langchain-google-genai==1.0.10
langchain==0.2.14
langchain-core==0.2.38
google-generativeai==0.7.2
transformers==4.44.2
torch==2.4.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
pydantic==2.9.2
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
textblob==0.18.0.post0
vaderSentiment==3.3.2
spacy==3.7.6
nltk==3.9.1
sentence-transformers==3.1.1
requests==2.32.3
pymongo==4.10.1
websockets==13.1
aioredis==2.0.1
python-multipart==0.0.12
bcrypt==4.2.0
passlib==1.7.4
python-jose==3.3.0
email-validator==2.2.0
aiohttp==3.10.8
asyncio==3.4.3